requirements.txt
# Core ML/DL Libraries
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
imgaug>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Framework
streamlit>=1.28.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
pandas>=2.0.0

# RAG Pipeline
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0
tiktoken>=0.5.0

# LLM Integration (Groq/Llama3)
requests>=2.31.0

# Optional: For Google Colab compatibility
# ipywidgets>=8.0.0
# jupyter>=1.0.0

